+++ src/index.ts
@@ -1,4 +1,7 @@
 export { default as ProSidebar } from './ProSidebar/ProSidebar';
+export { default as SidebarHeader } from './ProSidebar/SidebarHeader';
+export { default as SidebarContent } from './ProSidebar/SidebarContent';
+export { default as SidebarFooter } from './ProSidebar/SidebarFooter';
 export { default as Menu } from './ProSidebar/Menu';
 export { default as MenuItem } from './ProSidebar/MenuItem';
 export { default as SubMenu } from './ProSidebar/SubMenu';