+++ CHANGELOG.md
@@ -11,6 +11,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add suffix prop to menuItem and submenu
 - Add prefix prop to menuItem and submenu
+- Add sidebar layout
+
+  - Add sidebar header component
+  - Add sidebar content component
+  - Add sidebar footer component
 
 ## [0.2.0] - 2020-05-08